feat: Bootstrap Redux using redux,react-redux,react-actions,redux-persist,redux-thunk 😁 feat: Update the api(seeder,factory,migration) to serve multiple images of the product, provide lorem text for the product description fix: Use regex to match pathname for dynamic header title feat: Build ProductDetails component to present product data 😎 refactor: Rename Shop to ShopContainer refactor: Choose 2 sets of fonts(Serif, SanSerif) for app

+++ database/migrations/2019_05_10_180918_create_orders_table.php
@@ -16,8 +16,8 @@ class CreateOrdersTable extends Migration
         Schema::create('orders', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->timestamps();
-            $table->float('total');
-            $table->float('saved');
+            $table->decimal('total',12,2);
+            $table->decimal('saved',12,2);
             $table->boolean('tx_status')->nullable();
             $table->integer('tx_id')->nullable();
             $table->mediumText('err_mssg');
